Add the `name` property
Refactors the `src/data.json` to be an object instead of an array, this
way the keys are guaranteed to be unique. Separates the notion of the
directory vs name, now the directory (key in the top level object) can
be different than the name.

Ref. https://issues.redhat.com/browse/HACBS-2468