Do not set publicKey on RHTAP configs
The RHTAP configs have been updated to no longer specify the `publicKey`
attributes. This makes them more flexible by allowing them to be used in
different contexts.

The `update-infra-deployments.sh` script is updated to dynamically
inject the `publicKey` attribute when updating the ECP resources in the
infra-deployments repository. Since those resources are expected to be
used only in RHTAP, the publicKey value is still useful there.

Ref: EC-146

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>