Replace RHTAP with Konflux
For consistency I renamed the environments and corresponding template
files as well.

Note: For the "konflux-tasks" descriptions we talk about Red Hat
Trusted Tasks. I decided to just keep that since it still feels like
a reasonable title for the task definition policies, especially
since the task definitions they apply to currently are being used in
RHTAP/Dance as well as Konflux.

Ref: https://issues.redhat.com/browse/EC-378